Redo logout
