+++ web/packages/shared/storage/localForage.ts
@@ -11,3 +11,7 @@ if (haveWindow()) {
 }
 
 export default localForage;
+
+export const clearFiles = async () => {
+    await localForage.clear();
+};
