--- web/packages/shared/storage/localForage/helpers.ts
@@ -1,5 +1,0 @@
-import localForage from ".";
-
-export const clearFiles = async () => {
-    await localForage.clear();
-};
